# Reduce and update dependencies