--- a/src/SoapCore.Benchmark/SoapCore.Benchmark.csproj
+++ b/src/SoapCore.Benchmark/SoapCore.Benchmark.csproj
@@ -10,10 +10,10 @@
 	</PropertyGroup>
 
 	<ItemGroup Condition="$(TargetFramework) == 'netcoreapp3.1'">
-		<PackageReference Include="Microsoft.AspNetCore.TestHost" Version="3.1.0" />
+		<PackageReference Include="Microsoft.AspNetCore.TestHost" Version="3.1.32" />
 	</ItemGroup>
 	<ItemGroup Condition="$(TargetFramework) == 'net8.0'">
-		<PackageReference Include="Microsoft.AspNetCore.TestHost" Version="8.0.0" />
+		<PackageReference Include="Microsoft.AspNetCore.TestHost" Version="8.0.13" />
 	</ItemGroup>
 
 	<ItemGroup>