--- a/src/SoapCore/DocumentationWriter/SoapMethods.cs
+++ b/src/SoapCore/DocumentationWriter/SoapMethods.cs
@@ -1,9 +1,6 @@
-#if !NETCOREAPP3_1_OR_GREATER
-using Newtonsoft.Json;
-#endif
 using System.IO;
 using System.Linq;
-using System.Text;
+using System.Text.Json;
 using System.Xml;
 using System.Xml.Serialization;
 
@@ -60,11 +57,7 @@ namespace SoapCore.DocumentationWriter
 
 		public string GenerateDocumentation()
 		{
-#if NETCOREAPP3_1_OR_GREATER
-			var json = System.Text.Json.JsonSerializer.Serialize(this);
-#else
-			var json = JsonConvert.SerializeObject(this);
-#endif
+			var json = JsonSerializer.Serialize(this);
 
 			var html = $@"<!DOCTYPE html>
 <html lang=""en"">