--- a/src/SoapCore/SoapCore.csproj
+++ b/src/SoapCore/SoapCore.csproj
@@ -29,27 +29,21 @@
 	</ItemGroup>
 
 	<ItemGroup Condition="$(TargetFramework) == 'netstandard2.0' OR $(TargetFramework) == 'netstandard2.1'">
-		<PackageReference Include="Microsoft.AspNetCore.Http" Version="2.1.34" />
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.1.38" />
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.1.1" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="2.1.1" />
-		<PackageReference Include="System.IO.Pipelines" Version="4.7.0" />
-		<PackageReference Include="System.CodeDom" Version="4.7.0" />
-		<PackageReference Include="System.ServiceModel.Http" Version="4.8.1" />
-		<PackageReference Include="Microsoft.AspNetCore.Connections.Abstractions" Version="2.1.3" />
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.Abstractions" Version="2.1.38" />
+		<PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.3.0" />
+		<PackageReference Include="System.CodeDom" Version="9.0.2" />
+		<PackageReference Include="System.ServiceModel.Http" Version="8.1.1" />
+		<PackageReference Include="Microsoft.AspNetCore.Connections.Abstractions" Version="9.0.2" />
 		<PackageReference Include="System.Text.Json" Version="9.0.2" />
 	</ItemGroup>
 	<ItemGroup Condition="$(TargetFramework) == 'netcoreapp3.1'">
 		<FrameworkReference Include="Microsoft.AspNetCore.App" />
 		<PackageReference Include="System.CodeDom" Version="4.7.0" />
-		<PackageReference Include="System.ServiceModel.Http" Version="4.8.1" />
+		<PackageReference Include="System.ServiceModel.Http" Version="4.10.3" />
 	</ItemGroup>
 	<ItemGroup Condition="$(TargetFramework) == 'net8.0'">
 		<FrameworkReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Connections.Abstractions" Version="8.0.2" />
-		<PackageReference Include="System.CodeDom" Version="8.0.0" />
-		<PackageReference Include="System.ServiceModel.Http" Version="8.0.0" />
+		<PackageReference Include="System.CodeDom" Version="9.0.2" />
+		<PackageReference Include="System.ServiceModel.Http" Version="8.1.1" />
 	</ItemGroup>
 
 	<ItemGroup>
@@ -60,12 +54,9 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="System.Formats.Asn1" Version="8.0.1" />
-		<PackageReference Include="System.Security.Cryptography.Xml" Version="8.0.0" />
 		<PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="System.Text.Encodings.Web" Version="8.0.0" />
 	</ItemGroup>
 </Project>