With the last update of SoapCore 1.2.0 some dependencies were added that are not really necessary and were removed in previous versions (e.g. `System.Drawing.Common`).

- I have reduced the dependencies to the essentials and updated them.
- I have also replaced the dependency on `Newtonsoft.Json` with `System.Text.Json`.
- The tests failed for me due to the localization, because the tests are dependent on the text. I have therefore set the `CurrentUICulture ` for the relevant tests.
- In the readme it was listed that .NET 5.0 to 7.0 is supported, which is currently only the case with netstandard2.1. I have therefore updated this to .NET 8.0.